split concepts
