--- src/core/concepts.cxx
@@ -1,199 +1,0 @@
-module;
-#include "libfork/__impl/utils.hpp"
-export module libfork.core:concepts;
-
-import std;
-
-import libfork.utils;
-
-namespace lf {
-
-// ==== Returnable
-
-/**
- * @brief A type returnable from libfork's async functions/coroutines.
- *
- * This requires that `T` is `void` or a `std::movable` type.
- */
-export template <typename T>
-concept returnable = std::is_void_v<T> || (plain_object<T> && std::movable<T>);
-
-// ==== Stack
-
-template <typename T>
-  requires std::is_object_v<T>
-consteval auto constify(T &&x) noexcept -> std::add_const_t<T> &;
-
-/**
- * @brief Defines the API for a libfork compatible stack.
- *
- * // TODO: define if release is required before acquire?
- *
- * - After construction `this` is empty and push is valid.
- * - Pop is valid provided the FILO order is respected.
- * - Push produces pointers aligned to __STDCPP_DEFAULT_NEW_ALIGNMENT__.
- * - Destruction is expected to only occur when the stack is empty.
- * - Result of `.checkpoint()` is expected to:
- *     - Be "cheap to copy".
- *     - Have a null state (default constructed) that only compares equal to itself.
- *     - Is allowed to return null if push has never been called.
- *     - Compare equal if and only if they belong to the same stack or are both null.
- *     - Have no preconditions about when it's called.
- * - Prepare release puts the stack into a state which another thread can acquire it.
- * - Release detaches the current stack and leaves `this` empty.
- *     - This may be called concurrently with acquire
- * - Acquire attaches to the stack that the checkpoint came from:
- *     - It is only called the stack is empty.
- *     - It is only called with a checkpoint not equal to the current checkpoint.
- *     - It is called after prepare release (and no other functions in between)
- *
- * Fast-path operations: empty, push, pop, checkpoint
- * Slow-path operations: release, acquire
- */
-export template <typename T>
-concept worker_stack = plain_object<T> && requires (T stack, std::size_t n, void *ptr) {
-  { stack.push(n) } -> std::same_as<void *>;
-  { stack.pop(ptr, n) } noexcept -> std::same_as<void>;
-  { stack.checkpoint() } noexcept -> std::regular;
-  { stack.prepare_release() } noexcept -> std::movable;
-  { stack.release(stack.prepare_release()) } noexcept -> std::same_as<void>;
-  { stack.acquire(constify(stack.checkpoint())) } noexcept -> std::same_as<void>;
-};
-
-export template <typename T>
-concept aa_worker_stack = worker_stack<T> && true; // TODO: Allocator aware stack
-
-// ==== Context
-
-export template <typename T>
-struct steal_handle;
-
-export template <typename T>
-struct sched_handle;
-
-template <typename T>
-concept ref_to_worker_stack = std::is_lvalue_reference_v<T> && worker_stack<std::remove_reference_t<T>>;
-
-/**
- * @brief Defines the API for a libfork compatible worker context.
- *
- * This requires that `T` is an object type and supports the following operations:
- *
- * - Push/pop a frame handle onto the context in a LIFO manner.
- * - Have a `worker_stack` that can be accessed via `stack()`.
- * - Post an await handle to the context via `post()` and promise to call resume.
- */
-export template <typename T>
-concept worker_context =
-    plain_object<T> && requires (T context, steal_handle<T> frame, sched_handle<T> await) {
-      { context.push(frame) } -> std::same_as<void>;
-      { context.pop() } noexcept -> std::same_as<steal_handle<T>>;
-      { context.stack() } noexcept -> ref_to_worker_stack;
-    };
-
-/**
- * @brief Fetch the stack type of a worker context `T`.
- */
-export template <worker_context T>
-using stack_t = std::remove_reference_t<decltype(std::declval<T &>().stack())>;
-
-/**
- * @brief Fetch the checkpoint type of a worker context `T`.
- */
-export template <worker_context T>
-using checkpoint_t = decltype(std::declval<stack_t<T> &>().checkpoint());
-
-template <typename T>
-concept has_context_typedef = requires { typename std::remove_cvref_t<T>::context_type; };
-
-template <has_context_typedef T>
-using context_t = typename std::remove_cvref_t<T>::context_type;
-
-template <typename Derived, typename Base>
-concept derived_context_from = worker_context<Base> && std::derived_from<Derived, Base>;
-
-export template <typename Context>
-concept derived_worker_context =
-    has_context_typedef<Context> && derived_context_from<Context, context_t<Context>>;
-
-// ==== Scheduler
-
-/**
- * @brief An object capable of scheduling a libfork task for execution.
- *
- * These are typed to a context, the `post` method must:
- *
- * - Satisfy the strong exception guarantee.
- * - Guarantee eventual execution of the task associated with `handle`.
- */
-export template <typename Sch>
-concept scheduler =
-    has_context_typedef<Sch> && requires (Sch scheduler, sched_handle<context_t<Sch>> handle) {
-      { scheduler.post(handle) } -> std::same_as<void>;
-    };
-
-// ==== Forward-decl
-
-export template <worker_context>
-struct env {
-  explicit constexpr env(key_t) noexcept {}
-};
-
-export template <returnable T, worker_context Context>
-class task;
-
-// ========== Invocability ========== //
-
-template <typename Context>
-struct ctx_invoke_t {
-  // Explicitly constrained so overload resolution selects prefers
-  template <typename... Args, typename Fn>
-    requires std::invocable<Fn, env<Context>, Args...>
-  static constexpr auto operator()(Fn &&fn, Args &&...args)
-      LF_HOF(std::invoke(std::forward<Fn>(fn), env<Context>{key()}, std::forward<Args>(args)...))
-
-  template <typename... Args, typename Fn>
-  static constexpr auto operator()(Fn &&fn, Args &&...args)
-      LF_HOF(std::invoke(std::forward<Fn>(fn), std::forward<Args>(args)...))
-};
-
-template <typename Context, typename R>
-concept task_from = specialization_of<R, task> && std::same_as<Context, typename R::context_type>;
-
-/**
- * @brief Test if a callable `Fn` when invoked with `Args...` returns an `lf::task`.
- */
-export template <typename Fn, typename Context, typename... Args>
-concept async_invocable = worker_context<Context> &&                                                    //
-                          std::invocable<ctx_invoke_t<Context>, Fn, Args...> &&                         //
-                          task_from<Context, std::invoke_result_t<ctx_invoke_t<Context>, Fn, Args...>>; //
-
-/**
- * @brief Subsumes `async_invocable` and checks that the invocation is `noexcept`.
- */
-export template <typename Fn, typename Context, typename... Args>
-concept async_nothrow_invocable =
-    async_invocable<Fn, Context, Args...> && std::is_nothrow_invocable_v<ctx_invoke_t<Context>, Fn, Args...>;
-
-/**
- * @brief The result type of invoking an async function `Fn` with `Args...`.
- */
-export template <typename Fn, typename Context, typename... Args>
-  requires async_invocable<Fn, Context, Args...>
-using async_result_t = std::invoke_result_t<ctx_invoke_t<Context>, Fn, Args...>::value_type;
-
-/**
- * @brief Subsumes `async_invocable` and checks the result type is `R`.
- */
-export template <typename Fn, typename R, typename Context, typename... Args>
-concept async_invocable_to =
-    async_invocable<Fn, Context, Args...> && std::same_as<R, async_result_t<Fn, Context, Args...>>;
-
-/**
- * @brief Subsumes `async_nothrow_invocable` and `async_invocable_to`.
- */
-export template <typename Fn, typename R, typename Context, typename... Args>
-concept async_nothrow_invocable_to =
-    async_nothrow_invocable<Fn, Context, Args...> && async_invocable_to<Fn, R, Context, Args...>;
-
-} // namespace lf
